requirements.txt
# external requirements
importlib-metadata<5.0
importlib-resources==5.12.0
flake8
python-dotenv>=0.5.1
pytest
numpy
pandas>=1.3.0
seaborn
matplotlib
scipy
pytorch-lightning==2.4.0
torchmetrics==1.5.1
encodec==0.1.1
hydra-core==1.3.2
hydra_colorlog
scikit-learn
tables
ray[tune]

git+https://github.com/fadel/pytorch_ema
